+++ PolygonChecker/.vshistory/triangleSolver.h/2022-12-10_22_28_20_448.h
@@ -10,4 +10,4 @@ it will provide the user with the perimeter and area.*/
 
 char* analyzeTriangle(int side1, int side2, int side3);
 
-double* getTriangleAngles(double side1, double side2, double side3);
+int* getTriangleAngles(double side1, double side2, double side3);
